+++ src/features/logtalkTerminal.ts
@@ -7,6 +7,7 @@ import * as jsesc from "jsesc";
 import * as fs from "fs";
 import { spawn } from "process-promises";
 import LogtalkLinter from "./logtalkLinter";
+import LogtalkDeadCodeScanner from "./logtalkDeadCodeScanner";
 import { isFunction } from "util";
 import * as fsp from "fs/promises";
 import * as timers from "timers/promises";
@@ -326,30 +327,94 @@ export default class LogtalkTerminal {
     );
   }
 
-  public static async scanForDeadCode(uri: Uri) {
+  public static async scanForDeadCode(uri: Uri, deadCodeScanner: LogtalkDeadCodeScanner) {
     if (typeof uri === 'undefined') {
       uri = window.activeTextEditor.document.uri;
     }
+    let textDocument = null;
+    let logtalkHome: string = '';
+    let logtalkUser: string = '';
+    // Check for Configurations
+    let section = workspace.getConfiguration("logtalk");
+    if (section) { 
+      logtalkHome = jsesc(section.get<string>("home.path", "logtalk")); 
+      logtalkUser = jsesc(section.get<string>("user.path", "logtalk")); 
+    } else { 
+      throw new Error("configuration settings error: logtalk"); 
+    }
+    // Open the Text Document
+    await workspace.openTextDocument(uri).then((document: TextDocument) => { textDocument = document });
+    // Clear the Scratch Message File
+    let compilerMessagesFile  = `${logtalkUser}/scratch/.messages`;
+    await fsp.rm(`${compilerMessagesFile}`, { force: true });
+    // Create the Terminal
     LogtalkTerminal.createLogtalkTerm();
     const dir0: string = LogtalkTerminal.ensureDir(uri);
     const loader0 = path.join(dir0, "loader");
     const dir = path.resolve(dir0).split(path.sep).join("/");
     const loader = path.resolve(loader0).split(path.sep).join("/");
-    let goals = `logtalk_load(dead_code_scanner(loader)),logtalk_load('${loader}'),dead_code_scanner::directory('${dir}').\r`;
+    let goals = `vscode::dead_code('${dir}','${loader}').\r`;
     LogtalkTerminal.sendString(goals);
+    // Parse any compiler errors or warnings
+    const marker = path.join(dir0, ".vscode_dead_code_scanning_done");
+    await LogtalkTerminal.waitForFile(marker);
+    await fsp.rm(marker, { force: true });
+    if(fs.existsSync(`${compilerMessagesFile}`)) {
+      const lines = fs.readFileSync(`${compilerMessagesFile}`).toString().split(/\r?\n/);
+      let message = '';
+      for (const line of lines) {
+        message = message + line + '\n';
+        if(line == '*     ' || line == '!     ') {
+          deadCodeScanner.lint(textDocument, message);
+          message = '';
+        } 
+      }
+    }
   }
 
-  public static async rscanForDeadCode(uri: Uri) {
+  public static async rscanForDeadCode(uri: Uri, deadCodeScanner: LogtalkDeadCodeScanner) {
     if (typeof uri === 'undefined') {
       uri = window.activeTextEditor.document.uri;
     }
+    let textDocument = null;
+    let logtalkHome: string = '';
+    let logtalkUser: string = '';
+    // Check for Configurations
+    let section = workspace.getConfiguration("logtalk");
+    if (section) { 
+      logtalkHome = jsesc(section.get<string>("home.path", "logtalk")); 
+      logtalkUser = jsesc(section.get<string>("user.path", "logtalk")); 
+    } else { 
+      throw new Error("configuration settings error: logtalk"); 
+    }
+    // Open the Text Document
+    await workspace.openTextDocument(uri).then((document: TextDocument) => { textDocument = document });
+    // Clear the Scratch Message File
+    let compilerMessagesFile  = `${logtalkUser}/scratch/.messages`;
+    await fsp.rm(`${compilerMessagesFile}`, { force: true });
+    // Create the Terminal
     LogtalkTerminal.createLogtalkTerm();
     const dir0: string = LogtalkTerminal.getWorkspaceFolder(uri);
     const loader0 = path.join(dir0, "loader");
     const dir = path.resolve(dir0).split(path.sep).join("/");
     const loader = path.resolve(loader0).split(path.sep).join("/");
-    let goals = `logtalk_load(dead_code_scanner(loader)),logtalk_load('${loader}'),dead_code_scanner::rdirectory('${dir}').\r`;
+    let goals = `vscode::dead_code_recursive('${dir}','${loader}').\r`;
     LogtalkTerminal.sendString(goals);
+    // Parse any compiler errors or warnings
+    const marker = path.join(dir0, ".vscode_dead_code_scanning_done");
+    await LogtalkTerminal.waitForFile(marker);
+    await fsp.rm(marker, { force: true });
+    if(fs.existsSync(`${compilerMessagesFile}`)) {
+      const lines = fs.readFileSync(`${compilerMessagesFile}`).toString().split(/\r?\n/);
+      let message = '';
+      for (const line of lines) {
+        message = message + line + '\n';
+        if(line == '*     ' || line == '!     ') {
+          deadCodeScanner.lint(textDocument, message);
+          message = '';
+        } 
+      }
+    }
   }
 
   public static runTesters(uri: Uri) {