+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.21.0]
 
 * Add "Logtalk: Scan Dead Code (workspace)" command
+* Add `dead_code_scanner` tool warnings to the "Problems" pane
 * Update the "Known Issues" section in the readme file
 
 ## [0.20.0]